separated action-specific tools into droplet_actions_tools so the mapping between godo interface and tools file more clear. godo.DropletActionsService <--> DropletActionsTool godo.DropletService <--> DropletTool
