+++ internal/droplet/droplet_actions_tools_test.go
@@ -1,0 +1,193 @@
+package droplet
+
+import (
+	"context"
+	"encoding/json"
+	"errors"
+	"testing"
+
+	"github.com/digitalocean/godo"
+	"github.com/mark3labs/mcp-go/mcp"
+	"github.com/stretchr/testify/require"
+	"go.uber.org/mock/gomock"
+)
+
+func setupDropletActionsToolWithMocks(actions *MockDropletActionsService) *DropletActionsTool {
+	client := &godo.Client{}
+	client.DropletActions = actions
+	return NewDropletActionsTool(client)
+}
+
+func TestDropletActionsTool_powerCycleDroplet(t *testing.T) {
+	ctrl := gomock.NewController(t)
+	defer ctrl.Finish()
+
+	testAction := &godo.Action{ID: 111, Status: "in-progress"}
+	tests := []struct {
+		name        string
+		args        map[string]any
+		mockSetup   func(*MockDropletActionsService)
+		expectError bool
+	}{
+		{
+			name: "Successful power cycle",
+			args: map[string]any{"ID": float64(123)},
+			mockSetup: func(m *MockDropletActionsService) {
+				m.EXPECT().
+					PowerCycle(gomock.Any(), 123).
+					Return(testAction, nil, nil).
+					Times(1)
+			},
+		},
+		{
+			name: "API error",
+			args: map[string]any{"ID": float64(456)},
+			mockSetup: func(m *MockDropletActionsService) {
+				m.EXPECT().
+					PowerCycle(gomock.Any(), 456).
+					Return(nil, nil, errors.New("api error")).
+					Times(1)
+			},
+			expectError: true,
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			mockActions := NewMockDropletActionsService(ctrl)
+			if tc.mockSetup != nil {
+				tc.mockSetup(mockActions)
+			}
+			tool := setupDropletActionsToolWithMocks(mockActions)
+			req := mcp.CallToolRequest{Params: mcp.CallToolParams{Arguments: tc.args}}
+			resp, err := tool.powerCycleDroplet(context.Background(), req)
+			if tc.expectError {
+				require.NotNil(t, resp)
+				require.True(t, resp.IsError)
+				return
+			}
+			require.NoError(t, err)
+			require.NotNil(t, resp)
+			require.False(t, resp.IsError)
+			var outAction godo.Action
+			require.NoError(t, json.Unmarshal([]byte(resp.Content[0].(mcp.TextContent).Text), &outAction))
+			require.Equal(t, testAction.ID, outAction.ID)
+		})
+	}
+}
+
+func TestDropletActionsTool_powerOnDroplet(t *testing.T) {
+	ctrl := gomock.NewController(t)
+	defer ctrl.Finish()
+
+	testAction := &godo.Action{ID: 222, Status: "completed"}
+	tests := []struct {
+		name        string
+		args        map[string]any
+		mockSetup   func(*MockDropletActionsService)
+		expectError bool
+	}{
+		{
+			name: "Successful power on",
+			args: map[string]any{"ID": float64(123)},
+			mockSetup: func(m *MockDropletActionsService) {
+				m.EXPECT().
+					PowerOn(gomock.Any(), 123).
+					Return(testAction, nil, nil).
+					Times(1)
+			},
+		},
+		{
+			name: "API error",
+			args: map[string]any{"ID": float64(456)},
+			mockSetup: func(m *MockDropletActionsService) {
+				m.EXPECT().
+					PowerOn(gomock.Any(), 456).
+					Return(nil, nil, errors.New("api error")).
+					Times(1)
+			},
+			expectError: true,
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			mockActions := NewMockDropletActionsService(ctrl)
+			if tc.mockSetup != nil {
+				tc.mockSetup(mockActions)
+			}
+			tool := setupDropletActionsToolWithMocks(mockActions)
+			req := mcp.CallToolRequest{Params: mcp.CallToolParams{Arguments: tc.args}}
+			resp, err := tool.powerOnDroplet(context.Background(), req)
+			if tc.expectError {
+				require.NotNil(t, resp)
+				require.True(t, resp.IsError)
+				return
+			}
+			require.NoError(t, err)
+			require.NotNil(t, resp)
+			require.False(t, resp.IsError)
+			var outAction godo.Action
+			require.NoError(t, json.Unmarshal([]byte(resp.Content[0].(mcp.TextContent).Text), &outAction))
+			require.Equal(t, testAction.ID, outAction.ID)
+		})
+	}
+}
+
+func TestDropletActionsTool_powerOffDroplet(t *testing.T) {
+	ctrl := gomock.NewController(t)
+	defer ctrl.Finish()
+
+	testAction := &godo.Action{ID: 333, Status: "completed"}
+	tests := []struct {
+		name        string
+		args        map[string]any
+		mockSetup   func(*MockDropletActionsService)
+		expectError bool
+	}{
+		{
+			name: "Successful power off",
+			args: map[string]any{"ID": float64(123)},
+			mockSetup: func(m *MockDropletActionsService) {
+				m.EXPECT().
+					PowerOff(gomock.Any(), 123).
+					Return(testAction, nil, nil).
+					Times(1)
+			},
+		},
+		{
+			name: "API error",
+			args: map[string]any{"ID": float64(456)},
+			mockSetup: func(m *MockDropletActionsService) {
+				m.EXPECT().
+					PowerOff(gomock.Any(), 456).
+					Return(nil, nil, errors.New("api error")).
+					Times(1)
+			},
+			expectError: true,
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			mockActions := NewMockDropletActionsService(ctrl)
+			if tc.mockSetup != nil {
+				tc.mockSetup(mockActions)
+			}
+			tool := setupDropletActionsToolWithMocks(mockActions)
+			req := mcp.CallToolRequest{Params: mcp.CallToolParams{Arguments: tc.args}}
+			resp, err := tool.powerOffDroplet(context.Background(), req)
+			if tc.expectError {
+				require.NotNil(t, resp)
+				require.True(t, resp.IsError)
+				return
+			}
+			require.NoError(t, err)
+			require.NotNil(t, resp)
+			require.False(t, resp.IsError)
+			var outAction godo.Action
+			require.NoError(t, json.Unmarshal([]byte(resp.Content[0].(mcp.TextContent).Text), &outAction))
+			require.Equal(t, testAction.ID, outAction.ID)
+		})
+	}
+}
