+++ internal/droplet/droplet_tools_test.go
@@ -305,180 +305,3 @@ func TestDropletTool_deleteDroplet(t *testing.T) {
 		})
 	}
 }
-
-func TestDropletTool_powerCycleDroplet(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	testAction := &godo.Action{ID: 111, Status: "in-progress"}
-	tests := []struct {
-		name        string
-		args        map[string]any
-		mockSetup   func(*MockDropletActionsService)
-		expectError bool
-	}{
-		{
-			name: "Successful power cycle",
-			args: map[string]any{"ID": float64(123)},
-			mockSetup: func(m *MockDropletActionsService) {
-				m.EXPECT().
-					PowerCycle(gomock.Any(), 123).
-					Return(testAction, nil, nil).
-					Times(1)
-			},
-		},
-		{
-			name: "API error",
-			args: map[string]any{"ID": float64(456)},
-			mockSetup: func(m *MockDropletActionsService) {
-				m.EXPECT().
-					PowerCycle(gomock.Any(), 456).
-					Return(nil, nil, errors.New("api error")).
-					Times(1)
-			},
-			expectError: true,
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			mockDroplets := NewMockDropletsService(ctrl)
-			mockActions := NewMockDropletActionsService(ctrl)
-			if tc.mockSetup != nil {
-				tc.mockSetup(mockActions)
-			}
-			tool := setupDropletToolWithMocks(mockDroplets, mockActions)
-			req := mcp.CallToolRequest{Params: mcp.CallToolParams{Arguments: tc.args}}
-			resp, err := tool.powerCycleDroplet(context.Background(), req)
-			if tc.expectError {
-				require.NotNil(t, resp)
-				require.True(t, resp.IsError)
-				return
-			}
-			require.NoError(t, err)
-			require.NotNil(t, resp)
-			require.False(t, resp.IsError)
-			var outAction godo.Action
-			require.NoError(t, json.Unmarshal([]byte(resp.Content[0].(mcp.TextContent).Text), &outAction))
-			require.Equal(t, testAction.ID, outAction.ID)
-		})
-	}
-}
-
-func TestDropletTool_powerOnDroplet(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	testAction := &godo.Action{ID: 222, Status: "completed"}
-	tests := []struct {
-		name        string
-		args        map[string]any
-		mockSetup   func(*MockDropletActionsService)
-		expectError bool
-	}{
-		{
-			name: "Successful power on",
-			args: map[string]any{"ID": float64(123)},
-			mockSetup: func(m *MockDropletActionsService) {
-				m.EXPECT().
-					PowerOn(gomock.Any(), 123).
-					Return(testAction, nil, nil).
-					Times(1)
-			},
-		},
-		{
-			name: "API error",
-			args: map[string]any{"ID": float64(456)},
-			mockSetup: func(m *MockDropletActionsService) {
-				m.EXPECT().
-					PowerOn(gomock.Any(), 456).
-					Return(nil, nil, errors.New("api error")).
-					Times(1)
-			},
-			expectError: true,
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			mockDroplets := NewMockDropletsService(ctrl)
-			mockActions := NewMockDropletActionsService(ctrl)
-			if tc.mockSetup != nil {
-				tc.mockSetup(mockActions)
-			}
-			tool := setupDropletToolWithMocks(mockDroplets, mockActions)
-			req := mcp.CallToolRequest{Params: mcp.CallToolParams{Arguments: tc.args}}
-			resp, err := tool.powerOnDroplet(context.Background(), req)
-			if tc.expectError {
-				require.NotNil(t, resp)
-				require.True(t, resp.IsError)
-				return
-			}
-			require.NoError(t, err)
-			require.NotNil(t, resp)
-			require.False(t, resp.IsError)
-			var outAction godo.Action
-			require.NoError(t, json.Unmarshal([]byte(resp.Content[0].(mcp.TextContent).Text), &outAction))
-			require.Equal(t, testAction.ID, outAction.ID)
-		})
-	}
-}
-
-func TestDropletTool_powerOffDroplet(t *testing.T) {
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-
-	testAction := &godo.Action{ID: 333, Status: "completed"}
-	tests := []struct {
-		name        string
-		args        map[string]any
-		mockSetup   func(*MockDropletActionsService)
-		expectError bool
-	}{
-		{
-			name: "Successful power off",
-			args: map[string]any{"ID": float64(123)},
-			mockSetup: func(m *MockDropletActionsService) {
-				m.EXPECT().
-					PowerOff(gomock.Any(), 123).
-					Return(testAction, nil, nil).
-					Times(1)
-			},
-		},
-		{
-			name: "API error",
-			args: map[string]any{"ID": float64(456)},
-			mockSetup: func(m *MockDropletActionsService) {
-				m.EXPECT().
-					PowerOff(gomock.Any(), 456).
-					Return(nil, nil, errors.New("api error")).
-					Times(1)
-			},
-			expectError: true,
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			mockDroplets := NewMockDropletsService(ctrl)
-			mockActions := NewMockDropletActionsService(ctrl)
-			if tc.mockSetup != nil {
-				tc.mockSetup(mockActions)
-			}
-			tool := setupDropletToolWithMocks(mockDroplets, mockActions)
-			req := mcp.CallToolRequest{Params: mcp.CallToolParams{Arguments: tc.args}}
-			resp, err := tool.powerOffDroplet(context.Background(), req)
-			if tc.expectError {
-				require.NotNil(t, resp)
-				require.True(t, resp.IsError)
-				return
-			}
-			require.NoError(t, err)
-			require.NotNil(t, resp)
-			require.False(t, resp.IsError)
-			var outAction godo.Action
-			require.NoError(t, json.Unmarshal([]byte(resp.Content[0].(mcp.TextContent).Text), &outAction))
-			require.Equal(t, testAction.ID, outAction.ID)
-		})
-	}
-}
